Fixed SpringProb for Quartz; Added Functionality to add UserGroup

+++ dhis-mobile/dhis-service-mobile/src/main/java/org/hisp/dhis/mobile/impl/DefaultSendSMSService.java
@@ -50,7 +50,7 @@ public class DefaultSendSMSService implements SendSMSService
         return sendSMSStore.getAllSendSMS();
     }
     
-    public int getRowCount()
+    public long getRowCount()
     {
         return sendSMSStore.getRowCount();
     }
